[Refactor/#60] StaccatoDetail ViewModel 분리
