+++ Staccato-iOS/Staccato-iOS/Presentation/View/Home/GMSMapViewRepresentable.swift
@@ -81,9 +81,6 @@ extension GMSMapViewRepresentable.Coordinator: GMSMapViewDelegate {
     
     func mapView(_ mapView: GMSMapView, didTap marker: GMSMarker) -> Bool {
         if let userdata = marker.userData as? StaccatoCoordinateModel {
-            Task {
-                await viewModel.fetchStaccatoDetail(userdata.staccatoId)
-            }
             viewModel.modalNavigationState.navigate(to: .staccatoDetail(userdata.staccatoId))
         } else {
             print("⚠️ No StaccatoData found for this marker.")
